Move FreeRTOS hook functions from main to FreeRTOS folder
